BREAKING CHANGE: default version from V0 -> V1 (old names)
This commit also makes version and sequence available as optional fields to SMP Messages.  As before, if a header is not provided then it will be created, and now the version and sequence can be set more conveniently.  If a header is provided then it will take precedence for setting version and sequence.